stage1_pro_modular_training_system/stage1_finalV/requirements.txt
# ==============================================================================
# Stage-1 Pro System - PyTorch Lightning 2.4 Requirements (2025)
# ==============================================================================

# ==============================================================================
# CORE FRAMEWORKS (2025 Latest)
# ==============================================================================
torch==2.6.0
lightning==2.4.0
pytorch-lightning==2.4.0
transformers==4.56.0  # ✅ Official DINOv3 support (AutoModel, AutoImageProcessor)

# ==============================================================================
# CONFIGURATION
# ==============================================================================
hydra-core==1.4.0
hydra-zen==0.13.0
omegaconf==2.4.0
pydantic==2.9.0
pydantic-settings==2.5.0

# ==============================================================================
# DATA PROCESSING
# ==============================================================================
torchvision==0.20.0
timm==1.0.12
albumentations==1.4.24
pandas==2.2.3
numpy==2.2.0

# ==============================================================================
# MODEL OPTIMIZATIONS (2025)
# ==============================================================================
flash-attn==3.0.0

# ==============================================================================
# METRICS & EVALUATION
# ==============================================================================
torchmetrics==1.5.0
scikit-learn==1.6.0
scipy==1.14.0

# ==============================================================================
# CALIBRATION
# ==============================================================================
torchcp>=0.1.1

# ==============================================================================
# EXPERIMENT TRACKING
# ==============================================================================
wandb==0.18.0
mlflow==2.18.0

# ==============================================================================
# DATA VERSIONING
# ==============================================================================
dvc[gs,s3]==3.50.0

# ==============================================================================
# TESTING
# ==============================================================================
pytest==8.3.0
pytest-cov==6.0.0
pytest-mock==3.14.0
coverage==7.6.0

# ==============================================================================
# TYPE CHECKING
# ==============================================================================
pyright==1.1.370
mypy==1.13.0

# ==============================================================================
# CODE QUALITY
# ==============================================================================
ruff==0.6.0
black==24.1.0
isort==5.13.0
pre-commit==4.0.0

# ==============================================================================
# PRODUCTION (Deployment)
# ==============================================================================
onnx==1.18.0
onnxruntime-gpu==1.20.0

# ==============================================================================
# OPTIONAL: PROFILING (Install when needed)
# ==============================================================================
# nvidia-nsight-systems-cli==2024.5.1

# ==============================================================================
# OPTIONAL: GRADIENT-BASED TRAINING
# ==============================================================================
# sam-py>=1.0.0

# ==============================================================================
# OPTIONAL: MULTI-DATASET FUSION
# ==============================================================================
# pyarrow>=17.0.0
# opencv-python-headless>=4.10.0
